Struts webapp

+++ Last/src/com/cruds/leads/DeleteAction.java
@@ -11,6 +11,10 @@ import com.opensymphony.xwork2.ActionSupport;
 public class DeleteAction extends ActionSupport {
 	HttpServletRequest request = ServletActionContext.getRequest();
 	public String execute(){
+		if(request.getSession().getAttribute("isAdmin")==null)
+		{	
+			return ERROR;
+		}
 		int id = Integer.parseInt(request.getParameter("selectRecord"));
 		if(LeadDBUtil.deleteLead(id)) {
 			return SUCCESS;
